backend/requirements.txt
numpy==1.24.3
fastapi==0.121.2
keras==3.12.0
uvicorn==0.38.0
tensorflow==2.16.1
pillow==12.0.0
python-multipart==0.0.20